requirements.txt
streamlit
sqlalchemy
psycopg2-binary
openai
python-dotenv
pydantic
python-multipart
alembic
Office365-REST-Python-Client
boto3
PyPDF
pgvector